First crack at showing bf building rules. Some styling to fix
